Sleep fade: route audio through Web Audio GainNode for software volume
iOS Safari ignores writes to audio.volume (treats device volume as
hardware-only), so the JS-driven fade was always silent on the platform
the sleep timer matters on. Route the audio element through a Web Audio
GainNode the first time a sleep timer starts (a user gesture, required
for AudioContext.resume on iOS) and drive the fade via
gain.linearRampToValueAtTime, which runs in the audio thread.

Adds a 10-second sleep option for quick testing; parses data-minutes
with parseFloat so fractional values work.

+++ app.js
@@ -238,7 +238,7 @@ const App = {
             btn.addEventListener('click', e => {
                 e.stopPropagation();
                 const val = btn.dataset.minutes;
-                Player.startSleep(val === 'chapter' ? 'chapter' : parseInt(val));
+                Player.startSleep(val === 'chapter' ? 'chapter' : parseFloat(val));
                 document.getElementById('pp-sleep-menu').classList.remove('open');
             });
         });
@@ -277,7 +277,7 @@ const App = {
         document.querySelectorAll('#fs-sleep-menu button[data-minutes]').forEach(btn => {
             btn.addEventListener('click', () => {
                 const val = btn.dataset.minutes;
-                Player.startSleep(val === 'chapter' ? 'chapter' : parseInt(val));
+                Player.startSleep(val === 'chapter' ? 'chapter' : parseFloat(val));
                 document.getElementById('fs-sleep-menu').classList.remove('open');
             });
         });
